Removed person index service and moved all its methods to person service. Updated tests.
